requirement.txt
streamlit
requests
python-dotenv
sumy
gTTS
pandas
